+++ CloudHarvestCLI/__main__.py
@@ -1,7 +1,9 @@
 #!/usr/bin/env python
-from app import Harvest
+from CloudHarvestCLI.app import Harvest
 
 
 if __name__ == '__main__':
     with Harvest() as harvest:
         harvest.cmdloop()
+
+    print('Goodbye!')