+++ CloudHarvestCLI/commands/report/completers.py
@@ -1,11 +1,15 @@
+from logging import getLogger
 from typing import List
-from commands.arguments.completers import RemoteBaseCompleter
+from CloudHarvestCLI.commands.arguments.completers import RemoteBaseCompleter
 
+logger = getLogger('harvest')
 
 class ReportNameCompleter(RemoteBaseCompleter):
     def _run(self, *args, **kwargs) -> List[str]:
         try:
-            return [r['Name'] for r in self.result[0].get('data', [])]
+            return [r['Name'] for r in self.result['result']['data']]
 
         except Exception as ex:
+            logger.debug(ex)
+
             return []