+++ CloudHarvestCLI/commands/services/commandset.py
@@ -1,18 +1,18 @@
 from cmd2 import with_default_category, CommandSet, with_argparser, as_subcommand_to
-from processes import ConcurrentProcesses
-from .arguments import parser, services_attach_parser, services_kill_parser, services_list_parser
+from CloudHarvestCLI.processes import ConcurrentProcesses
+from CloudHarvestCLI.commands.services.arguments import parser, services_attach_parser, services_kill_parser, services_list_parser
 
 
 @with_default_category('Harvest')
 class ServicesCommand(CommandSet):
     @with_argparser(parser)
     def do_services(self, args):
-        from commands.base import get_subtask
+        from CloudHarvestCLI.commands.base import get_subtask
         get_subtask(parent=self, parser=parser, args=args)
 
     @as_subcommand_to('services', 'attach', services_attach_parser, help='Attach a progress bar to a running job.')
     def attach(self, args):
-        from text.printing import print_message
+        from CloudHarvestCLI.messages import print_message
 
         for process in ConcurrentProcesses.objects:
             if process.name == args.name and hasattr(process, 'attach_progressbar'):
@@ -25,7 +25,7 @@ class ServicesCommand(CommandSet):
 
     @as_subcommand_to('services', 'kill', services_kill_parser, help='Stop jobs.')
     def kill(self, args):
-        from text.printing import print_message
+        from CloudHarvestCLI.messages import print_message
 
         for process in ConcurrentProcesses.objects:
             if process.name in args.names and hasattr(process, 'kill'):
@@ -39,7 +39,7 @@ class ServicesCommand(CommandSet):
 
     @as_subcommand_to('services', 'list', services_list_parser, help='Get a list of running jobs')
     def list(self, args):
-        from text.printing import print_data
+        from CloudHarvestCLI.text.printing import print_data
 
         keys, data = ConcurrentProcesses().report()
         print_data(data=data, keys=keys)