+++ CloudHarvestCLI/exceptions.py
@@ -1,0 +1,5 @@
+from CloudHarvestCoreTasks.tasks.base import BaseHarvestException
+
+class HarvestClientException(BaseHarvestException):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)